requirements.txt
python-docx
python-dotenv
pydantic
pytest
black
isort
mypy
langchain
langchain-core
langchain-groq
loguru
streamlit
PyPDF2
typing-extensions
tqdm
aiohttp